requirements/base.txt
appnope==0.1.0
boto==2.38.0
decorator==4.0.6
dj-database-url==0.3.0
Django==1.8.8
django-braces==1.8.1
django-crispy-forms==1.5.2
django-filter==0.11.0
django-grappelli==2.7.2
django-model-utils==2.4
django-nose==1.4.2
django-oauth-toolkit==0.10.0
django-pnm==0.1.6
django-storages==1.1.8
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
ecdsa==0.13
Fabric==1.10.2
fabtools==0.19.0
factory==1.2
factory-boy==2.6.0
funcsigs==0.4
geopy==1.11.0
gnureadline==6.3.3
gunicorn==19.4.0
httplib2==0.9.2
httpretty==0.8.12
ipython==4.0.1
ipython-genutils==0.1.0
Jinja2==2.8
MarkupSafe==0.23
mercurial==3.6.2
mock==1.3.0
msgpack-python==0.4.6
nose==1.3.7
oauth2==1.9.0.post1
oauthlib==1.0.3
paramiko==1.16.0
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.5
Pillow==2.9.0
psycopg2==2.6.1
ptyprocess==0.5
pycrypto==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
PyYAML==3.11
redis==2.10.5
requests==2.9.0
requests-oauthlib==0.6.0
responses==0.5.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
whitenoise==2.0.6
supervisor==3.2.1